copy: split copy_nowindows.go
