+++ copy/copy_unix.go
@@ -8,7 +8,6 @@ import (
 	"syscall"
 
 	"github.com/containerd/containerd/sys"
-	"github.com/containerd/continuity/sysx"
 	"github.com/pkg/errors"
 	"golang.org/x/sys/unix"
 )
@@ -43,24 +42,6 @@ func copyFileContent(dst, src *os.File) error {
 	bufferPool.Put(buf)
 
 	return err
-}
-
-func copyXAttrs(dst, src string) error {
-	xattrKeys, err := sysx.LListxattr(src)
-	if err != nil {
-		return errors.Wrapf(err, "failed to list xattrs on %s", src)
-	}
-	for _, xattr := range xattrKeys {
-		data, err := sysx.LGetxattr(src, xattr)
-		if err != nil {
-			return errors.Wrapf(err, "failed to get xattr %q on %s", xattr, src)
-		}
-		if err := sysx.LSetxattr(dst, xattr, data, 0); err != nil {
-			return errors.Wrapf(err, "failed to set xattr %q on %s", xattr, dst)
-		}
-	}
-
-	return nil
 }
 
 func copyDevice(dst string, fi os.FileInfo) error {
